Refactor Substance export workflow
